Xjw 0922

See merge request td_group/linggongyun_v3!30

+++ merchant/src/main/java/com/example/merchant/service/impl/MenuServiceImpl.java
@@ -14,6 +14,7 @@ import com.example.mybatis.vo.QueryPassRolemenuVO;
 import com.example.mybatis.vo.RoleMenuPassVO;
 import com.example.mybatis.vo.RoleMenuVO;
 import com.example.redis.dao.RedisDao;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -154,18 +155,17 @@ public class MenuServiceImpl extends ServiceImpl<MenuDao, Menu> implements MenuS
     @Override
     @Transactional(rollbackFor = Exception.class)
     public ReturnJson savePlatRole(SaveManagersRoleDTO saveManagersRoleDto, String managersId) {
-
         Managers managersOne = managersDao.selectOne(new QueryWrapper<Managers>().eq("mobile_code",
                 saveManagersRoleDto.getMobileCode()));
-        if (managersOne != null) {
-            return ReturnJson.error("此手机号码已近注册过！");
-        }
         Managers managers = new Managers();
-        BeanUtils.copyProperties(saveManagersRoleDto, managers);
-        managers.setParentId(managersId);
-        managers.setUserSign(3);
-        managers.setPassWord(PWD_KEY + MD5.md5(saveManagersRoleDto.getPassWord()));
+
         if ("".equals(saveManagersRoleDto.getId())) {
+            BeanUtils.copyProperties(saveManagersRoleDto, managers);
+            managers.setParentId(managersId);
+            managers.setUserSign(3);
+            if (managersOne != null) {
+                return ReturnJson.error("此手机号码已近注册过！");
+            }
             managersDao.insert(managers);
             String[] menuId = saveManagersRoleDto.getMenuIds().split(",");
             for (int i = 0; i < menuId.length; i++) {
@@ -176,6 +176,13 @@ public class MenuServiceImpl extends ServiceImpl<MenuDao, Menu> implements MenuS
             }
             return ReturnJson.success("添加成功");
         } else {
+            managers = managersDao.selectById(saveManagersRoleDto.getId());
+            if (StringUtils.isNotEmpty(saveManagersRoleDto.getPassWord())) {
+                managers.setPassWord(PWD_KEY + MD5.md5(saveManagersRoleDto.getPassWord()));
+            }
+            if (!managersOne.getId().equals(managers.getId())) {
+                return ReturnJson.error("此手机号码已近注册过！");
+            }
             managersDao.updateById(managers);
             objectMenuDao.delete(new QueryWrapper<ObjectMenu>().eq("object_user_id", managers.getId()));
             String[] menuId = saveManagersRoleDto.getMenuIds().split(",");
